requirements.txt
requests
speedtest-cli
icmplib